requirements.txt
pandas~=1.2.4
scikit-learn~=0.24.2
PyYAML~=5.4.1
matplotlib~=3.4.2
numpy~=1.20.3
shap~=0.39.0
seaborn~=0.11.2